Shifted mock session to new file for use in different test files
